Set up an abstract interface for backends and moved the original backend to `vosk_local`
